requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.4.0
appnope==0.1.4
assemblyai==0.37.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.13.12
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
environs==11.0.0
executing==2.1.0
fastapi==0.112.2
fastapi-cli==0.0.5
fastapi_cors==0.0.6
fastavro==1.10.0
filelock==3.17.0
Flask==3.0.0
flatbuffers==25.2.10
fsspec==2025.2.0
gast==0.4.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.17.0
google-pasta==0.2.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.8
inflection==0.5.1
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.13.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.4.0
narwhals==1.26.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.63.0
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
Quandl==3.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.5.1
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
streamlit==1.42.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.1
termcolor==2.5.0
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.3
typer==0.12.5
types-requests==2.32.0.20241016
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.1
wrapt==1.17.2
